pyproject.toml
[project]
name = "uv-fastapi-template-base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosmtplib>=4.0.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.1",
    "celery>=5.4.0",
    "celery-aio-pool>=0.1.0rc8",
    "cryptography>=44.0.2",
    "fastapi>=0.115.6",
    "google-api-python-client>=2.162.0",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.5",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.37",
    "sqlalchemy-utils>=0.41.2",
    "uvicorn[standard]>=0.34.0",
    "uvloop>=0.21.0",
]

[dependency-groups]
dev = [
    "diagrams>=0.24.4",
    "faker>=33.3.1",
    "freezegun>=1.5.1",
    "httpx>=0.28.1",
    "mypy>=1.14.1",
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.1",
    "pytest-cov>=6.0.0",
    "readmeai[google-generativeai]>=0.6.1",
    "respx>=0.22.0",
    "ruff>=0.8.5",
]

[tool.pytest.ini_options]
addopts = "-p no:warnings --cov --cov-fail-under=50 --cov-report=term-missing"
# log_cli = true
pythonpath = "./src"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 135
src = ["src", "tests"]
exclude = [".venv", "migrations"]

[tool.ruff.lint]
ignore = [
    "RUF012",
]

select = [
    "F",
    "E",
    "I",
]
[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
strict = true
ignore_missing_imports = true
disallow_untyped_defs = false
